ips: isp: bug [x2j2-1284] edit size type
Summary: Signed-off-by: yongyong.duan <yongyong.duan@horizon.ai>

Test Plan: no

Reviewers: zhongkai.yin, xiaofeng.ling

Reviewed By: zhongkai.yin, xiaofeng.ling

Subscribers: yongyong.duan

Differential Revision: https://cr.hobot.cc/D3029

+++ drivers/ips/isp/x2/isp_dev.h
@@ -21,13 +21,13 @@
 #define ISP_NAME    "x2-isp"
 
 struct isp_dev_s {
-	/* if ipu mem from ion */
+	/* if isp mem from ion */
 	struct ion_client *isp_iclient;
 	struct ion_handle *isp_ihandle;
-	/* if ipu mem from reserve mem */
+	/* if isp mem from reserve mem */
 	phys_addr_t mapbase;
 	phys_addr_t mapbaseio;
-	uint32_t memsize;
+	size_t memsize;
 	void __iomem *regbase;
 	void *vaddr;
 	int irq;
